feat(tools): add password strength indicator

+++ core/model/src/commonMain/kotlin/com/thejohnsondev/model/tools/PasswordStrength.kt
@@ -1,6 +1,6 @@
 package com.thejohnsondev.model.tools
 
 data class PasswordStrength(
-    val level: Int,
+    val level: Float,
     val suggestion: String?
 )
